[tests] Split one big test file into multiple files covering smaller areas
